Don't create files in /dev
This addresses the case when the user specifies that a file in /dev
should be written to, but for whatever reason, the SDCard is missing.
The result was that a normal file was created and things got weird when
the real SDCard was inserted again.

This changes prevents fwup from creating files in /dev. A clever user
can get around this through cleverly crafted paths, but accidental file
creation shouldn't be possible.

Fixes https://github.com/fhunleth/fwup/issues/35.