Remove conditional when computing progress
This conditional was triggering a coverity error, but it also got messed
up during the progress computation refactoring. Even though it seems
harmless to skip progress computation when not outputing progress, since
this case gets less testing than the progress-printing cases, don't skip
it. Anyway, the progress computation is instantaneous compared to SDCard
programming, so this is useless optimization.