Handle archive_read_data_block returning 0, but not EOF
This seems to occur with a very unlucky alignment of bytes in the
archive. It was produced once in a Nerves .fw file. I have not seen it
anywhere else on any other project and it appears to be coupled to the
libarchive version since it wouldn't reproduce with the desktop version
(non-Buildroot made) version of fwup.