Add require-path-on-device feature
This is useful for embedded systems that have no easy way of querying
filesystem metadata for which partition is the one being used. This goes
directly to the OS to see which it is. Arguably, this is a more robust
way of solving this problem that what had been done in the past.