Fix stdin being read past end
When streaming, libarchive would keep reading data from stdin for as
long as it would come. For big .fw files being streamed over a net
connection, this could take a while if you messed on part of a command
or if there was an error. This change fixes the issue and adds a unit
test for it.