Don't advertise fat_rm!()
The semantics for fat_rm!() are backwards when compared to fat_rm().
I.e., fat_rm!() is the one that fails if the file doesn't exist whereas
fat_rm() succeeds whether or not the file exists. This seems backwards
to me, but is only this way for backwards compatibility. This may be
something to fix for fwup 1.0, but for now, leave it undocumented.