Fail when files are missing from the archive
Previously, it wasn't guaranteed that all on-resource handlers would be
fired. Normally this was the case, but there was a use case where it was
convenient to add or remove a file to trigger different behavior. This
use case can be implemented other ways. Verifying that all on-resource
handlers are run is so important in ensuring that updates are completely
applied and is pretty much expected behavior from anyone using fwup.
This removes that misfeature.