New fwup buffering properly merges partial blocks so no-zeroing needed
The new caching is smart enough to merge blocks instead of zeroing
across a hole. This means that the code in the unit test that zeros
things to block boundaries should be deleted.

It would be an improvement to implement holes as trim operations (not
necessarily sending hardware TRIM commands) since the normal semantics
of holes is that they contain zeros. That means that the
read/modify/write merge that the caching algorithm is doing for this
case should really be changed to skip the read operation.