Fix trim boundary to be after the last byte
This fixes an issue where if a Flash device or file was not a multiple
of 128 KiB, the very last bytes would be internally trimmed by fwup and
ignored. This could cause data loss if it ever happened and something
was stored in the last bytes. It was found as a side effect of the media
size testing in the next commit.