Fix block_cache init for devices w/o lseek end support
This fixes an issue on OSX where destination SDCards were assumed to be
blank since lseek SEEK_END returned 0, and the code assumed that meant
the whole thing was blank.

This manifested itself in "upgrade" tasks not working on OSX. This
was not a common use case, but it was reall surprising when it didn't
work.