Add path_write, pipe_write, and execute
These make it possible to use fwup on raw NAND systems using UBI. Rather
than pulling the UBI tools into fwup, commands are added to shell out to
existing utilities. While this is not in the spirit of fwup, it will
hopefully help gain some real-world use to better guide what's needed.
All of the commands are experimental and can only be used with the
--unsafe flag.