Fix FAT overwrite error; simplify truncation
This fixes an error that you'd get if you tried to write over a file
that already existed on the FAT partition. It also simplifies the
file truncation logic to not need to always rm files before writing to
them.

Added unit tests to make sure these don't break again in the future.