Reduce progress bar width to fit thin terminals
It doesn't look like it's possible to query the terminal width and
adjust automatically. The use case that brought this issue to light was
where fwup was being run over ssh disconnected from the terminal and
stdin was used to stream the update. That prevents using VT100 queries
to get around the terminal disconnect. I'm not sure that there's another
way, so I reduced the progress bar width for everyone. This seems ok
since it's basically eye candy anyway.

Fixes #77.