Skip OS-dependent sparse file tests on APFS
This fixes #81. The problem is that hole creation on APFS seems to vary
between runs. On Linux it is predictable and the regression test was
easy to write. Without this predicatability, some other way of running
the tests would have to be figured out.

I know of only a couple uses of sparse file support in the wild and the
APFS behavior only causes those cases to be slightly slower in their
updates. I can't think of a situation where the APFS behavior would
break anything. If anything, the situation is slightly better since OSX
didn't support sparse files before High Sierra.

The non-fs dependent sparse file support code should verify that things
work internally on OSX, so I'm not worried about regressions from
commenting this out.