Resolve paths relative to the conf file
This fixes an issue where it was difficult to specify relative paths in
conf files since the resolution was relative to the current working
directory. This changes resolution to be relative to the conf file. This
makes it possible to remove a lot of environment variables in .conf
files whose sole purpose was to force absolute paths.