Force timestamps for FAT operations
Previously the .fw file creation time (meta-creation-time) was used for
the timestamp for all FAT filesystem operations. This could lead to
non-bit-for-bit compatibility between .fw files with identical contents.
The way around this issue was to set $NOW to force timestamps.  I don't
think that anyone actually does this in practice. Now that UUID support
was added and it doesn't take into account the creation time, the FAT
operations need to be forced to a consistent time. For lack of a better
timestamp, Jan 1, 1980 is used since no one will mistake it for a real
timestamp.