Fail if root tries to write to a read-only file
While this is normally allowed, it's almost certainly a mistake for fwup
users. The use case for writing to regular files with fwup is to create
images of SDCard/eMMC memory. It's trivial to set file permissions on
those.

This fixes an error found when packaging fwup on Alpine.  Test 68 tests
writing to a read-only file and expects it to fail.  The Alpine packager
uses fakeroot so doesn't fail like expected since file permissions are
ignored.