Fix segfault when verifying on OSX Catalina
Something has changed on OSX Catalina where allocating a 4K buffer on the
stack causes a segfault now. Reducing the size to 2K works, but it
seemed safer to allocate the buffer on the heap and not worry about the
size. Since it's on the heap, increase the size since there's plenty of
memory.