Account for 128 KB writes for final size
If the destination is not a multiple of 128 KB, it's possible to write
beyond the end since fwup writes in 128 KB blocks. This reduces the
destination length to the nearest multiple so that that can't happen.

This fixes an issue where writing the secondary GPT partition would fail
due to the added padding in the final block. Unfortunately, for drives
that are not a multiple of 128 KB, the secondary GPT is non-compliant
since it's not at the last block. I would assume that whatever logic
recovers from a corrupt primary GPT won't work since the GPT header
won't be at the expected LBA.

This should be fixed. However, as a minor consolation, manufactoring
Flash programmers typically take image files as input where the size of
the image is much less than the destination Flash size. The secondary
GPT header is going to be placed incorrectly by design and if it's
important to update it, then that will need to be done at runtime.