Add support for writing dmcrypt'd partitions
This only adds support for the aes-cbc-plain mode for dmcrypt.

The raw_write function now supports optional named parameters for
specifying a cipher and secret. The only cipher is "aes-cbc-plain" which
is sufficient for experimenting with encrypted partitions with
device-specific secrets.

See the README.md additions and new unit tests for trivial examples. The
unit tests hardcode passwords for convenience. Real use should not do
this.