Use SOURCE_DATE_EPOCH in metadata check tests
This fixes an issue that was introduce when fwup started preferring
SOURCE_DATE_EPOCH over NOW for making deterministic archives. Metadata
tests had used NOW, but that was ignored on build systems that were
actually using SOURCE_DATE_EPOCH.

The fix is to temporarily override SOURCE_DATE_EPOCH so that the unit
tests work. This verifies that SOURCE_DATE_EPOCH works. Additionally,
009_metadata_cmdline.test has an additional test now that verifies that
the old NOW variable continues to work.

Fixes #113.