Support building minimal versions of fwup
This change makes it possible to trim down the footprint of fwup. This
primarily removes the capability of creating firmware archives since
that is almost never used on devices.

To use:

```sh
./configure --enable-minimal-build.
```

Note that the regression tests will fail since archives can't be
created. The solution to this is to build a full-featured and minimal
version of fwup and use the full-featured one to create the archives.
See the `build_and_test_minimal.sh` script for details.