Add support for applying xdelta3-generated patches
This adds support for applying firmware updates where binaries inside of
them have been replaced with xdelta3 patch files. This makes it possible
for firmware update servers (or users can do this by hand) to make
create `.fw` files that reuse data that's already on target devices.

For example, it's common for people using fwup to have A/B partitions on
their devices and copy down the full root filesystem. The idea here is
that the server can know what bits are on the device if the device does
not modify its root filesystem and reports which firmware UUID is
running. The firmware update server then uses then xdelta3 program to
generate a patch file from going from the current firmware to the new
firmware. That patch file should be much smaller to send down. By
replacing the full rootfs with the patch file, the device's version of
`fwup` can process the patch. `fwup` still verifies that the end result
matches expected Blake2B hash and cryptographic signatures still work
since they're used to check the final outcome rather than the input.