Change the skip-holes default to off
Fwup has a feature where it can respect sparse file holes and recreate
them on devices. This can majorly save time writing large ext2/4
partitions to Flash. It has a couple gotchas that make it tricky to use:

1. If programs creating the sparse files create holes to initialize runs
   of zeros, those holes may not contain zeros on the device. It would be
   nice if the programs wrote zeros, but that can't be enforced.
2. Sparse hole sizes and offsets depend on the filesystems where the
   archives are created. This introduces another variation when trying
   make reproducible builds.
3. Sparse files are incompatible with the upcoming binary delta feature.
   In other words, any files that are stored as sparse ones cannot be
   turned into patches.

Since sparse file reproduction is an optimization, it was decided to
disable it by default. This will prevent it from accidentally getting
used and preventing binary delta support.

Overall, this feature turned out to not be used as much as anticipated
since so many fwup users are not writing large ext2/4 partitions to
their devices. Users who do this can still take advantage of sparse
files by adding "skip-hole = true" to the file resources of interest.