Refactor verify to match apply decode structure
This reduces the chance of false positives/negatives due to subtleties
in libarchive processing between versions. It fixes a false verify
failure on Ubuntu 14.04 (Travis) that doesn't affect applies.

Overall, though, it does seem better that the verify code path exercises
as much similar code as the apply path since that's what we really want
to work and is why we run verify.