Verify writes to device files
There is a not-well understood issue in one deployment where `fwup` can
successfully write to Flash memory but when checked, there was
corruption detected on some devices in the fleet. The Flash memory can
be written again and the issue will go away. There are enough unknowns,
potentially low quality parts in use, and environmental factors that
this might not be an issue with fwup at all.

Prior to this patch, `fwup` would not read back what it had written.
This patch changes this behavior to verify writes to device files. The
`--verify-writes` and `--no-verify-writes` commandline options can be
used explicitly enable or disable this feature. Writes are not verified
to regular files by default.

The way this works is that every time `fwup` writes a block of data, it
will read it back and compare it. `fwup` already bypasses OS caching so
that it can enforce a deterministic write order to media and this
feature plugs into that code. `fwup` writes are always 128 KB.

Simple experiments that the overhead from verifying data is hard to see
when sending updates over a network. Firmware updates that are applied
locally do seem a little slower. The rationale for making the default to
verify writes for raw device files is that it is the safer option and
the performance impact is small.