No writes after a failure
Previously, there was an attempt to flush inflight writes when errors
occur. This stops this from happening. I.e., any error will stop all
writes that fwup still has not issued to the OS.

The reason for this change is that if a write error occurs, the call to
flush could result in another error that would be reported to users
instead of the original one. The first one was closer to the root cause
and more helpful when debugging.

On error handlers can still try to write. This is important since they
may need to report an issue or try to undo something. Of course, they
can't count on anything actually being written due to the cache. This
has always been the case, but now even more won't make it to disk on
error.

This fixes an issue with the unit test where it depended on earlier
writes making it to "disk" in the error handler. While previously this would always
work in this specific unit test, it shouldn't be the case in general and
was a bad example. This moves the part in questions (a FAT mkfs) to
another run of fwup.

Finally, I have a different opinion now with fwup and writes after
errors.  I believe that my previous logic for flushing is wrong, and the
simplest course of action is for fwup to stop as soon as an
unrecoverable error has been hit and minimize writes.