Reduce window for FAT corruption
The FatFs has recommendations for reducing the chance of FAT
datastructure corruption on unexpected termination. See
http://elm-chan.org/fsw/ff/doc/appnote.html#critical

This is a bandaid since making FATFS resiliant to unexpected power
outages does not look possible. fwup's block-based caching (it only writes
128 KB blocks at a time) confounds this. At the very least, though, I
believe that this change reduces the amount of time that the fwup cache
is out of sync with fatfs, AND that is enough of an improvment to
warrant calling f_sync after every FatFs write operation.