Fix autodetection of built-in SDCard readers
The built-in SDCard device driver in Linux marks these drives a
non-removal. See the comment at
https://elixir.bootlin.com/linux/v5.8.18/source/drivers/mmc/core/block.c#L2335.

This change modifies the drive detection logic to consider all built-in
SDCard readers to hold removable block devices from `fwup`'s point of
view. Even if the Linux kernel maintainers view these devices as
non-removable, this change should be safe to nearly all `fwup` users
since it's rare to use an SDCard for a system drive. There's another
check, though, so if you use an SDCard for a root filesystem, `fwup`
will not autodetect it.