fatfs: bump from v0.14a to v0.14b
Here are the release notes:

```
R0.14b (April 17, 2021)
  Made FatFs uses standard library <string.h> for copy, compare and search instead of built-in string functions.
  Added support for long long integer and floating point to f_printf(). (FF_STRF_LLI and FF_STRF_FP)
  Made path name parser ignores the terminating separator to allow "dir/".
  Improved the compatibility in Unix style path name feature.
  Fixed the file gets dead-locked when f_open() failed with some conditions. (appeared at R0.12a)
  Fixed f_mkfs() can create wrong exFAT volume due to a timing dependent error. (appeared at R0.12)
  Fixed code page 855 cannot be set by f_setcp(). (appeared at R0.13)
  Fixed some compiler warnings.
```

The most visible change to fwup was that the FAT serial numbers all
changed. In 0.14a, the FAT serial number was calculated from the current
time. Fwup forces the time for repeatable builds, so the serial numbers
were all 0021-0000. The serial number is now the current time plus the
volue size. That doesn't seem like a problematic change (could even be
desirable), so all of the fwup unit tests were updated.