Allow specifying a fallback task
In very specific cases, it might be desirable to fallback to another
task when the target task fails. Most commonly this will be falling
back to the `complete` task when the `upgrade` fails due to the firmware
not having a previous firmware.

The fallback will only occur if the failure is withing the tasks functions
so checks before and after the task do not trigger a fallback (i.e. issues
with archive or config file)

This is inspired by nerves-project/nerves#682