Use resource name when fat_write's filename isn't specified
This is needed to support the file resources being automatically added.
The nice part is that since it creates backwards compatible archives, it
can be used as a convenience in newer fwup.conf files.

For example, before you'd have to write:

```
        on-resource my_file.bin {
                fat_write(\${BOOT_PART_OFFSET}, "my_file.bin")
        }
```

Now you only have to specify `my_file.bin` once:

```
        on-resource my_file.bin {
                fat_write(\${BOOT_PART_OFFSET})
        }
```

This works for directories too so long as the resource has the directory
name in it.