Remove image file before corruption tests
This isn't a problem now, but the upcoming potential change to remove
redundent writes actually avoids corruption completely since the writes
that would be corrupt get skipped. This is cool, but defeats the purpose
of the corruption tests. This update starts everything fresh for each
test which is cleaner anyway.