Add --minimize-writes option
This option currently defaults to off.

By passing --minimize-writes, fwup will read a block before writing it.
If the block is unchanged, the write is skipped entirely. This is
desirable if you're really interested in minimizing write operations to
Flash memory. Reasons include Flash being slow and it's common to be
rewriting similar data, worry about wearing Flash out prematurely, and
general worry that each write has a small chance to fail.