Support partial final writes and add --max-size option
There was a constraint that would round the destination to a multiple of
128KB segments so that <128KB writes didn't need to be handled. These
are now handled, but only for the very last 128KB segment.

A --max-size option was added to make testing this feature possible on
regular files. For real devices, the max size is automatically
determined from the device. This option can reduce the max size of a
real device.

Internally, the variable tracking the max size (end_offset) is now a
hard limit. It used to be a soft limit so that it was possible to infer
a good guess from a regular file. That was interesting for testing MBR
and GPT functionality, but not really anywhere else. The hard limit
makes it possible calculate the partial write size at the end of the
device and to give a better error if fwup is told to write past the end.