Improve GPT test with partial final segment
This adjusts the generation script to add some padding that's not in any
partition to push the secondary header into a partial final segment.
This is a better test since it exercises fwup's GPT secondary header
placement code in a more obvious way. The test came too close to working
without the partial final segment commit for comfort.