Fix redundant U-Boot env selection on wrap
An incorrect U-Boot environment was being selected at the 128th
generation. Thanks to @RossiSoftLtd for identifying the problem and
figuring out how to fix it.

This also adds a unit test to verify that the right environment is
selected after running fwup enough times to wrap the 8-bit counter.

Fixes #214