Remove escaping of backslashes to support literal ${var}
THIS IS A BACKWARDS INCOMPATIBLE CHANGE THAT AFFECTS FW FILE CREATION.
It does not affect processing of fw files so you don't need to worry
about change fwup versions on existing devices or worry about applying
.fw files created with old versions of fwup.

Automatic escaping of backslashes made it impossible to write `${var}`
to a U-Boot variable. Here's why. When doing this, you have to remember
that `fwup` evaluates variable substitution twice - once when making the
.fw file and once when applying it. You obviously have to escape the `$`
when creating the .fw file. That made sense. Then to survive the apply
step, you'd think that you could double escape the `$`. You'd be wrong,
though, since `fwup` was escaping the backslashes that you were adding
writing the configuration to the .fw file. Therefore, the variable
substitution was guaranteed to happen since you couldn't double escape a
`$`.

Amazingly, the auto-escaping behavior was only tested in the regression
tests via the exec test on Windows - a combination that would never be
used for real.

When you run into this issue, it's weird enough to be pretty confusing,
imho. Hopefully that and how rare should have been in real uses cases
makes this something that no one actually did.

This changes string processing to not automatically escape backslashes
when creating the .fw file. It is possible to escape a `$` through to
the end. This allows you to write a U-Boot environment variable with a
`${var}` in it. This also locks down string processing behaviors by
adding unit tests.