When writing regular files, check file size to support expansion
This is the way that programming external media works. It was never
hooked up for regular files even though this is a very convenient use
case when using Qemu. This change to implements it.

The way this works is that Qemu users first create a disk image. This is
just a big sparse file and then they run fwup to program the initial
image. The disk image size will be total image size in this case, so the
expansion and end-seeking functionality is expected to work.

When the destination file does not exist or is 0, fwup works as it
previously does by only writing the minimum size possible or following
the `--max-size` commandline argument.

IMPORTANT: Regular files no longer automatically expand if they're
nonzero in size. This can break test scripts, but if it were still
supported, forcing a max size like is done for physical media wouldn't
work for qemu disk images. The fix is to remove old image files if you
don't care about them or to use `dd seek=x` to make them the right size
or to update the complete task in the fwup files to write to the full
size. The last option is what's being done for the unit tests.