Extend fat_cp to copy between FAT filesystems
The use case that motivated this was the need to copy the Raspberry Pi
bootloader to two partitions when using the autoboot feature. In this
case the bootloader blob needs to be on the first partition (autoboot)
and also on whatever next partition that the autoboot.txt points. Due to
the streaming nature of fwup, the way to get one file to two places is
to save it in one place and then copy it to the second.