Change docs for streaming WN message to indicate info
The use of this message for actual warnings is so rare that saying it is
a warning is misleading. This message comes from calls to info() in
fwup.conf files nearly all of the time.

This changes the docs, but not the letters used for the message so
there's no concern about backwards compatibility.