Fix hang when flushing the write cache on error
This addresses the problem in issue #253 that caused fwup to hang rather
than return an error. It shouldn't have returned an error, but hanging
was worse.

This also adds the .fw files that triggered the issue since they seem to
be 100% reliable in reproducing the issue whereas my other attempts were
not.