Support soft limits for regular files
Regular files worked differently depending on whether fwup was run to
create an image file or to modify an image file. When creating the file,
the fwup instructions could expand it as much as disk space allowed.
However, when modifying the file, fwup limited the max length to the
existing lenght or whatever was passed on the commandline.

This commit adds the concept of a soft limit when modifying existing
files. This soft limit can be bumped when writing, but gets used when
expanding partitions to fill space.