Return smaller chunks at a time with deltas
This removes the constraint where progress reporting was based on
xdelta3's window size configuration. By default, this was 8 MB which
resulted in 8 MB being returned on every call. Since 8 MB actually took
quick a bit of time to accumulate and write, it led to the progress bar
sticking for 30-60 seconds or more. This was quite unsettling to say the
least.

This change returns a maximum of 128 KB chunks at a time which matches
other write buffering sizes, so it hopefully is a happy medium of
allowing efficient write block sizes while allowing the code to return
to do the writes and report progress.