Remove quotes on meta.conf strings when unneeded
This removes a backwards/forwards compatibility issue with fwup. The
problem is that libconfuse doesn't support 64-bit integers so fwup uses
strings to hold them. This primarily affects block offsets since they
can exceed 32-bits pretty easily.

libconfuse can parse strings with or without double quotes. However, it
can only parse integers without double quotes. For example, if fwup
needs to start storing an integer as a string and then writes that
integer as a string to the meta.conf, only versions of fwup that support
that field as a string will be able to parse it. On the other hand, if
fwup stores it in the meta.conf without double quotes, old and new
versions of fwup can parse it.

The easiest way to apply this rule is to generically dequote all strings
that look like they don't need double quotes. Note that if you have a
real string that looks like an integer, it's still ok, since it was
optional to add double quotes.

This has a side effect of causing a change in output between fwup
versions. Everything is still compatible, but the same input to fwup
will give a different output because of the removed double quotes. This
is very easy to detect if looking at firmware UUIDs. Sadly, if this
change didn't cause it to happen, a likely follow up one would, so I
think we just have to live with it. Concerned users probably lock fwup
versions anyway.

The check for whether a string needs double quotes or not is probably
too conservative. I wanted to get rid of a lot of double quotes to take
the hit once. However, some things like hyphens made me worry about
accidentally affecting a math expression in the future.