Add a simple two-word nickname for UUIDs
This addresses an issue where different firmware files are mistakenly
thought to be the same despite having different UUIDs. These are almost
always situations where the firmware version numbers weren't bumped
since the firmware was in actively being developed or debugged.

This commit adds the concept of a firmware nickname that's derived from
the UUID. Since it's derived from the UUID, no new metadata is stored.
In fact, the algorithm could be copied out of fwup to avoid the upgrade.

This nickname is two words where the words are chosen from a 256-word
list. That means they only have 16-bits of uniqueness compared to the
124-bits in the UUID. For the intended use case, it should be highly
unlikely for a firmware update to go unnoticed in development when
watching nicknames.

Word list storage is a little more than 1KB, so the feature is a trivial
add to fwup's footprint.